renamed reviewsList.js and reviewsMetadata.js to dummyReviewsData.js and dummyReviewsMetadata.js
